requirements.txt
asttokens==2.4.0
backcall==0.2.0
black==23.9.1
blinker==1.6.2
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.5
comm==0.1.4
contourpy==1.1.1
cycler==0.10.0
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.4
Flask==2.3.3
fonttools==4.42.1
idna==2.10
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==2.1.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.16.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.6
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
supervision==0.14.0
sympy==1.12
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.180
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.3.7
